+++ test/resources/expiry-doc-definitions.js
@@ -1,0 +1,44 @@
+function() {
+  function dynamicExpiryType(doc) {
+    if (typeof doc.expiry === 'string') {
+      return 'datetime';
+    } else if (typeof doc.expiry === 'number') {
+      return 'integer';
+    } else {
+      return 'object';
+    }
+  }
+
+  return {
+    staticNumberExpiryDoc: {
+      typeFilter: simpleTypeFilter,
+      channels: { write: 'write' },
+      propertyValidators: { },
+      expiry: 4077433455
+    },
+    staticDateStringExpiryDoc: {
+      typeFilter: simpleTypeFilter,
+      channels: { write: 'write' },
+      propertyValidators: { },
+      expiry: '2050-12-31T24:00Z'
+    },
+    staticDateExpiryDoc: {
+      typeFilter: simpleTypeFilter,
+      channels: { write: 'write' },
+      propertyValidators: { },
+      expiry: new Date(Date.UTC(3018, 3, 15, 3, 26, 58))
+    },
+    dynamicExpiryDoc: {
+      typeFilter: simpleTypeFilter,
+      channels: { write: 'write' },
+      propertyValidators: function(doc, oldDoc) {
+        return {
+          expiry: { type: dynamicExpiryType(doc) }
+        };
+      },
+      expiry: function(doc, oldDoc) {
+        return doc.expiry;
+      }
+    }
+  };
+}